% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.r
\name{ros_categories}
\alias{ros_categories}
\title{Rosette API categorizatioin service}
\usage{
ros_categories(content, genre = NULL, language = NULL)
}
\arguments{
\item{content}{either a URI or character vector of content to process}

\item{genre}{document genre (optional)}

\item{language}{Language: ISO 639-3 code (optional)}
}
\description{
Returns a category identified in the input. The possible categories are Tier 1
contextual categories defined in the QAG Taxonomy The response is the contextual
categories identified in the input.
}
\examples{
txt <- c('Sony Pictures is planning to shoot a good portion of the new "Ghostbusters" in Boston as well.')
ros_categories(txt)
}

