% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.r
\name{ros_embedding}
\alias{ros_embedding}
\title{Rosette API text embedding service}
\usage{
ros_embedding(content, genre = NULL, language = NULL)
}
\arguments{
\item{content}{either a URI or character vector of content to process}

\item{genre}{document genre (optional)}

\item{language}{Language: ISO 639-3 code (optional)}
}
\description{
Provides the embedding for the input text. Returns a single vector of floating point
numbers for your input, known as a text embedding. Among other uses, a text embedding
enables you to calculate the similarity between two documents or two words.
}
\examples{
ros_embedding("Cambridge, Massachusetts")
}

