% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.r
\name{ros_entities}
\alias{ros_entities}
\title{Rosette API entity extraction service}
\usage{
ros_entities(content, genre = NULL, language = NULL)
}
\arguments{
\item{content}{either a URI or character vector of content to process}

\item{genre}{document genre (optional)}

\item{language}{Language: ISO 639-3 code (optional)}
}
\description{
Returns each entity extracted from the input. The response is a list of extracted
entities. Each entity includes an entityId (Wikidata QID or temporary ID), mention
(entity text in the input), normalized text (complete form of the entity in the input),
and the count (number of mentions of entity in the input).
}
\examples{
txt <- c("Bill Murray will appear in new Ghostbusters film: Dr. Peter Venkman was ",
         "spotted filming a cameo in Boston this… http://dlvr.it/BnsFfS")
txt <- paste0(txt, collapse="")
ros_entities(txt)
}

