% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.r
\name{ros_language}
\alias{ros_language}
\title{Rosette API language identification service}
\usage{
ros_language(content, genre = NULL, language = NULL)
}
\arguments{
\item{content}{either a URI or character vector of content to process}

\item{genre}{document genre (optional)}

\item{language}{Language: ISO 639-3 code (optional)}
}
\description{
Returns list of candidate languages in order of descending confidence. The response is
an ordered list of identified languages, including language and identification confidence,
sorted by descending confidence.
}
\examples{
ros_language("Por favor Señorita, says the man.")
}

