% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.r
\name{ros_make_name}
\alias{ros_make_name}
\title{Make a Name object}
\usage{
ros_make_name(name, entity_type = NULL, language = NULL, script = NULL)
}
\arguments{
\item{name}{the name to be matched}

\item{entity_type}{the entity type of the text. One of \code{PERSON}, \code{LOCATION} or
\code{ORGANIZATION} (optional)}

\item{language}{ISO 639-3 code for the name's langauge (optional)}

\item{script}{ISO 15924 code for the name's script (optional)}
}
\description{
Make a Name object
}
\examples{
ros_make_name("Michael Jackson")
}

