% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.r
\name{ros_name_similarity}
\alias{ros_name_similarity}
\title{Rosette API version info}
\usage{
ros_name_similarity(name1, name2)
}
\arguments{
\item{name1, name2}{a name object constructed by a call to \code{ros_make_name()}}
}
\description{
Returns the confidence score of matching 2 names. You must specify the names to be matched.
}
\examples{
ros_name_similarity(ros_make_name("Michael Jackson"), ros_make_name("迈克尔·杰克逊"))
}

