% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.r
\name{ros_name_translation}
\alias{ros_name_translation}
\title{Rosette API name translation service}
\usage{
ros_name_translation(name, target_language, target_scheme = NULL,
  target_script = NULL, entity_type = NULL, source_language_origin = NULL,
  source_language_use = NULL, source_script = NULL)
}
\arguments{
\item{name}{string (name) to be translated}

\item{target_language}{ISO 639-3 code for the translation langauge}

\item{target_scheme}{transliteration scheme for the translation (optional)}

\item{target_script}{ISO 15924 code for the translation script (optional)}

\item{entity_type}{the entity type of the text. One of \code{PERSON}, \code{LOCATION} or
\code{ORGANIZATION} (optional)}

\item{source_language_origin}{ISO 639-3 code for the name's language of origin (optional)}

\item{source_language_use}{ISO 639-3 code for the name's language of use (optional)}

\item{source_script}{ISO 15924 code for the name's script (optional)}
}
\description{
Returns the translation of a name. You must specify the name to translate and the
target language for the translation.
}
\examples{
ros_name_translation("معمر محمد أبو منيار القذاف", "eng")
}

