% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.r
\name{ros_relationships}
\alias{ros_relationships}
\title{Rosette API relationship extraction service}
\usage{
ros_relationships(content, genre = NULL, language = NULL)
}
\arguments{
\item{content}{either a URI or character vector of content to process}

\item{genre}{document genre (optional)}

\item{language}{Language: ISO 639-3 code (optional)}
}
\description{
Returns each relationship extracted from the input. The response is a list of
extracted relationships.
}
\examples{
ros_relationships("The Ghostbusters movie was filmed in Boston.")
}

