% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentences.r
\name{ros_sentences}
\alias{ros_sentences}
\title{Rosette API sentence determination service}
\usage{
ros_sentences(content, genre = NULL, language = NULL)
}
\arguments{
\item{content}{either a URI or character vector of content to process}

\item{genre}{document genre (optional)}

\item{language}{Language: ISO 639-3 code (optional)}
}
\description{
Divides the input into sentences. The response contains a list of sentences.
}
\examples{
lines <- c("This land is your land. This land is my land",
           "From California to the New York island;",
           "From the red wood forest to the Gulf Stream waters",
           "",  "This land was made for you and Me.",
           "", "As I was walking that ribbon of highway,",
           "I saw above me that endless skyway:",
           "I saw below me that golden valley:",
           "This land was made for you and me.")
ros_sentences(paste0(lines, collapse="\\\\n"))
}

