% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.r
\name{ros_sentiment}
\alias{ros_sentiment}
\title{Rosette API sentiment analysis service}
\usage{
ros_sentiment(content, genre = NULL, language = NULL)
}
\arguments{
\item{content}{either a URI or character vector of content to process}

\item{genre}{document genre (optional)}

\item{language}{Language: ISO 639-3 code (optional)}
}
\description{
Analyzes the positive and negative sentiment expressed by the input. The response
contains sentiment analysis results.
}
\examples{
txt <- c("Original Ghostbuster Dan Aykroyd, who also co-wrote the 1984 Ghostbusters ",
         "film, couldn’t be more pleased with the new all-female Ghostbusters cast, ",
         "telling The Hollywood Reporter, “The Aykroyd family is delighted by this ",
         "inheritance of the Ghostbusters torch by these most magnificent women in comedy.")
ros_sentiment(paste0(txt, collapse=""))
}

