% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.r
\name{ros_tokens}
\alias{ros_tokens}
\title{Rosette API tokenizatioin service}
\usage{
ros_tokens(content, genre = NULL, language = NULL)
}
\arguments{
\item{content}{either a URI or character vector of content to process}

\item{genre}{document genre (optional)}

\item{language}{Language: ISO 639-3 code (optional)}
}
\description{
Divides the input into tokens. The response contains a list of tokens.
}
\examples{
ros_tokens("北京大学生物系主任办公室内部会议")
}

