% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.r
\name{rosette_api_key}
\alias{rosette_api_key}
\title{Get or set ROSETTE_API_KEY value}
\usage{
rosette_api_key(force = FALSE)
}
\arguments{
\item{force}{force setting a new Rosette API key for the current environment?}
}
\value{
atomic character vector containing the Rosette API key
}
\description{
The API wrapper functions in this package all rely on a Rosette API
key residing in the environment variable \code{ROSETTE_API_KEY}. The
easiest way to accomplish this is to set it in the `\code{.Renviron}` file in your
home directory.
}

